+++ lib/config/micro-main.d.ts
@@ -1,0 +1,5 @@
+export declare const getConfig: (env: string) => any;
+export declare const deps: {
+    'webpack-config-single-spa-ts': string;
+    'ts-config-single-spa': string;
+};