+++ lib/config/paths.d.ts
@@ -1,0 +1,9 @@
+export declare const paths: {
+    src: string;
+    dist: string;
+    public: string;
+    js: string;
+    css: string;
+    img: string;
+    assets: string;
+};