+++ lib/index.d.ts
@@ -1,12 +1,14 @@
 import { Factory } from 'fbi';
 import CommandBuild from './commands/build';
 import CommandServe from './commands/serve';
+import CommandInstall from './commands/install';
+import TemplateVue from './templates/vue';
 import TemplateReact from './templates/react';
-import TemplateMicro from './templates/micro';
+import TemplateMicroMain from './templates/micro-main';
 export default class FactoryWeb extends Factory {
     id: string;
     description: string;
-    commands: (CommandServe | CommandBuild)[];
-    templates: (TemplateReact | TemplateMicro)[];
+    commands: (CommandServe | CommandInstall | CommandBuild)[];
+    templates: (TemplateVue | TemplateReact | TemplateMicroMain)[];
     execOpts: any;
 }