+++ lib/templates/react.d.ts
@@ -1,16 +1,12 @@
-import { Template } from 'fbi';
-import * as ejs from 'ejs';
-import Factory from '../index';
-export default class TemplateReact extends Template {
+import Factory from '..';
+import BaseClass from './base';
+export default class TemplateReact extends BaseClass {
     factory: Factory;
     id: string;
-    description: string;
     path: string;
-    renderer: typeof ejs.render;
+    description: string;
     templates: never[];
     constructor(factory: Factory);
     protected gathering(flags: Record<string, any>): Promise<void>;
     protected writing(): Promise<void>;
-    protected installing(flags: Record<string, any>): Promise<void>;
-    protected ending(): Promise<void>;
 }