+++ src/commands/build.ts
@@ -1,10 +1,9 @@
-import { join } from 'path'
-import { Command } from 'fbi'
+import type { Configuration } from 'webpack'
 
 import Factory from '..'
-import { REACT_TEMPLATE_ID, MICRO_TEMPLATE_ID } from '../const'
-
-const runReactBuildScript = require('./react/scripts/build.js')
+import { Command } from 'fbi'
+import webpack from 'webpack'
+import webpackConfig from '../config'
 
 export default class CommandBuild extends Command {
   id = 'build'
@@ -33,34 +32,56 @@ export default class CommandBuild extends Command {
       unknown
     )
 
-    const _cwd = process.cwd()
-    const distDirName = 'dist'
-    const distDir = join(_cwd, distDirName)
-    // const srcDir = join(_cwd, tsconifg.compilerOptions.rootDir||'src')
-
     this.logStart(`Start building:`)
-    this.logItem(`remove '${distDirName}'...`)
-    await this.fs.remove(distDir)
+    const template = this.context.get('config.factory.template')
+    const config = webpackConfig(template, {
+      env: process.env.NODE_ENV
+    })
 
-    const execOpts: any = {
-      ...this.factory.execOpts,
-      stdio: flags.debug ? 'inherit' : 'pipe',
-      env: {
-        BUILD_ENV: flags.mode ?? 'development'
-      }
-    }
-    const templateId = this.context.get('config.factory.template')
     try {
-      const buildSpinner = this.createSpinner(`building...`).start()
-      if (templateId === REACT_TEMPLATE_ID) {
-        await runReactBuildScript()
-      } else if (templateId === MICRO_TEMPLATE_ID) {
-        await this.exec.command('umi build', execOpts)
-        buildSpinner.succeed()
-      }
+      await this.build(config)
     } catch (err) {
       this.error('Failed to build project')
-      this.error(err).exit()
+      // this.error(err).exit()
+      console.log(err)
+      this.exit()
     }
+  }
+
+  protected build(config: Configuration) {
+    const compiler = webpack(config)
+
+    return new Promise((resolve, reject) => {
+      compiler.run((err, stats) => {
+        if (err) {
+          console.log('---')
+          console.error(err.stack || err)
+          if ((err as any).details) {
+            console.error((err as any).details)
+          }
+          reject()
+        }
+
+        console.log(
+          stats?.toString({
+            chunks: false, // Makes the build much quieter
+            colors: true // Shows colors in the console
+          })
+        )
+
+        // const info = stats?.toJson()
+
+        // if (stats?.hasErrors()) {
+        //   console.error(info.errors)
+        //   reject()
+        // }
+
+        // if (stats?.hasWarnings()) {
+        //   console.warn(info.warnings)
+        // }
+
+        resolve()
+      })
+    })
   }
 }