+++ src/index.ts
@@ -1,15 +1,19 @@
 import { join } from 'path'
 import { Factory } from 'fbi'
+
 import CommandBuild from './commands/build'
 import CommandServe from './commands/serve'
+import CommandInstall from './commands/install'
+
+import TemplateVue from './templates/vue'
 import TemplateReact from './templates/react'
-import TemplateMicro from './templates/micro'
+import TemplateMicroMain from './templates/micro-main'
 
 export default class FactoryWeb extends Factory {
   id = 'factory-web'
-  description = 'factory for web factory development'
-  commands = [new CommandBuild(this), new CommandServe(this)]
-  templates = [new TemplateMicro(this), new TemplateReact(this)]
+  description = 'factory for web application development'
+  commands = [new CommandBuild(this), new CommandServe(this), new CommandInstall(this)]
+  templates = [new TemplateVue(this), new TemplateReact(this), new TemplateMicroMain(this)]
 
   execOpts: any = {
     cwd: process.cwd(),