+++ templates/micro-main/src/config/dev.config.ts
@@ -1,0 +1,3 @@
+import { IEnvConfig } from './index'
+
+export default {} as IEnvConfig