+++ templates/react/src/config/dev.config.ts
@@ -1,3 +1,0 @@
-export default {
-  host: 'https://48p1r2roz4.sse.codesandbox.io',
-}