+++ templates/vue/.eslintrc.js
@@ -1,0 +1,5 @@
+const { vueEslint } = require('fbi-lint')
+
+module.exports = {
+  ...vueEslint
+}