+++ templates/vue3/src/source.d.ts
@@ -1,2 +1,0 @@
-declare const React: string
-declare module '*.png'